// src/schema.d.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

/**
 * A complete description of a musical temperament with additional metadata.
 */
export interface Temperament {
  /**
   * The name of the temperament.
   */
  name: string;
  /**
   * A description of the temperament.
   */
  description?: string;
  /**
   * The source from which the temperament data was obtained.
   */
  source?: string;
  /**
   * The name of the base note of the octave.
   */
  octaveBaseName: string;
  /**
   * The frequency (in Hz) of the reference pitch.
   */
  referencePitch: number;
  /**
   * The note name of the reference pitch.
   */
  referenceName: string;
  /**
   * The octave number of the reference pitch.
   */
  referenceOctave: number;
  /**
   * A mapping of note names to a description of each note.
   */
  notes: {
    /**
     * A description of the note, as an offset from another note.
     */
    [k: string]: [string, number];
  };
  [k: string]: any;
}